requirements.txt
async-generator==1.10
async-timeout==3.0.1
cycler==0.10.0
discord==1.7.3
discord.py==1.7.3
PyGithub==1.55
pypi==2.1
python-dateutil==2.8.1
python-dotenv==0.19.2
requests==2.27.1
